name/empty_records is useful for the GoogleFonts spec as well
(issue #2238)
