+++ tests/specifications/adobe_fonts_test.py
@@ -5,27 +5,6 @@ from fontbakery.checkrunner import (PASS, FAIL)
 from fontbakery.utils import TEST_FILE
 
 
-def test_check_name_empty_records():
-    from fontbakery.specifications.adobe_fonts import (
-        com_adobe_fonts_check_name_empty_records as check)
-    font_path = TEST_FILE("source-sans-pro/OTF/SourceSansPro-Regular.otf")
-    test_font = TTFont(font_path)
-
-    # try a font with fully populated name records
-    status, message = list(check(test_font))[-1]
-    assert status == PASS
-
-    # now try a completely empty string
-    test_font['name'].names[3].string = b''
-    status, message = list(check(test_font))[-1]
-    assert status == FAIL
-
-    # now try a string that only has whitespace
-    test_font['name'].names[3].string = b' '
-    status, message = list(check(test_font))[-1]
-    assert status == FAIL
-
-
 def test_check_consistent_upm():
     from fontbakery.specifications.adobe_fonts import (
         com_adobe_fonts_check_consistent_upm as check)
